test: centralise process test helpers
